Add parsing of modsettings.lsx + tests
