+++ tests/Games/NexusMods.Games.Larian.Tests/BaldursGate3/BG3ModsettingsFileFormatTests.SerializeModsettingsLoadOrder_VerifyTest.verified.txt
@@ -23,22 +23,6 @@
               <attribute id="Version64" type="int64" value="1" />
             </node>
             <node id="ModuleShortDesc">
-              <attribute id="Folder" type="LSString" value="Carry Weight Increased 9000" />
-              <attribute id="MD5" type="LSString" value="" />
-              <attribute id="Name" type="LSString" value="Carry Weight Increased 9000" />
-              <attribute id="PublishHandle" type="uint64" value="" />
-              <attribute id="UUID" type="guid" value="92af0ebd-a04c-4e9c-8306-b947aac4c127" />
-              <attribute id="Version64" type="int64" value="1" />
-            </node>
-            <node id="ModuleShortDesc">
-              <attribute id="Folder" type="LSString" value="Dnd5rBardCollegeOfSwordsHomebrew_fb035e66-c916-8b33-a4cd-af40b86e4813" />
-              <attribute id="MD5" type="LSString" value="92ae5c2df98777c464300fb9c7a5e6d2" />
-              <attribute id="Name" type="LSString" value="DnD 5R Bard - College of Swords (Homebrew)" />
-              <attribute id="PublishHandle" type="uint64" value="4343434" />
-              <attribute id="UUID" type="guid" value="fb035e66-c916-8b33-a4cd-af40b86e4813" />
-              <attribute id="Version64" type="int64" value="36028797018963969" />
-            </node>
-            <node id="ModuleShortDesc">
               <attribute id="Folder" type="LSString" value="MetamagicExtendedQuickened2SP" />
               <attribute id="MD5" type="LSString" value="" />
               <attribute id="Name" type="LSString" value="MetamagicExtendedQuickened2SP" />
@@ -55,12 +39,28 @@
               <attribute id="Version64" type="int64" value="36028799166447617" />
             </node>
             <node id="ModuleShortDesc">
+              <attribute id="Folder" type="LSString" value="Carry Weight Increased 9000" />
+              <attribute id="MD5" type="LSString" value="" />
+              <attribute id="Name" type="LSString" value="Carry Weight Increased 9000" />
+              <attribute id="PublishHandle" type="uint64" value="" />
+              <attribute id="UUID" type="guid" value="92af0ebd-a04c-4e9c-8306-b947aac4c127" />
+              <attribute id="Version64" type="int64" value="1" />
+            </node>
+            <node id="ModuleShortDesc">
               <attribute id="Folder" type="LSString" value="WaypointInsideEmeraldGrove" />
               <attribute id="MD5" type="LSString" value="" />
               <attribute id="Name" type="LSString" value="Waypoint Inside Emerald Grove" />
               <attribute id="PublishHandle" type="uint64" value="" />
               <attribute id="UUID" type="guid" value="e342ee75-f7c9-4aeb-b6de-403991578337" />
               <attribute id="Version64" type="int64" value="72057598332895232" />
+            </node>
+            <node id="ModuleShortDesc">
+              <attribute id="Folder" type="LSString" value="Dnd5rBardCollegeOfSwordsHomebrew_fb035e66-c916-8b33-a4cd-af40b86e4813" />
+              <attribute id="MD5" type="LSString" value="92ae5c2df98777c464300fb9c7a5e6d2" />
+              <attribute id="Name" type="LSString" value="DnD 5R Bard - College of Swords (Homebrew)" />
+              <attribute id="PublishHandle" type="uint64" value="4343434" />
+              <attribute id="UUID" type="guid" value="fb035e66-c916-8b33-a4cd-af40b86e4813" />
+              <attribute id="Version64" type="int64" value="36028797018963969" />
             </node>
           </children>
         </node>
